Harden pnpm split-store installs
