+++ nix/workspace-tools/lib/mk-bun-cli/bun-deps.nix
@@ -68,6 +68,11 @@ else
       else
         export PNPM_HOME="$PWD/pnpm-home"
         export PNPM_STORE_DIR="$PWD/pnpm-store"
+        export npm_config_store_dir="$PNPM_STORE_DIR"
+        export npm_config_manage_package_manager_versions=false
+        export npm_config_side_effects_cache=false
+        export npm_config_verify_store_integrity=true
+        export npm_config_package_import_method=copy
         export NPM_CONFIG_NODE_LINKER=hoisted
       fi
 
@@ -100,7 +105,13 @@ else
               --frozen-lockfile \
               --force \
               --shamefully-hoist \
-              --ignore-scripts
+              --ignore-scripts \
+              --config.confirmModulesPurge=false \
+              --config.manage-package-manager-versions=false \
+              --config.side-effects-cache=false \
+              --config.verify-store-integrity=true \
+              --config.package-import-method=copy \
+              --config.store-dir="$PNPM_STORE_DIR"
           ) 2>&1 | tee "$install_log"
         fi
         if [ "''${PIPESTATUS[0]:-0}" -ne 0 ]; then
